+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- `Innmind\Http\Headers::all()`
+
 ## 6.2.0 - 2023-04-02
 
 ### Added